moved model evaluation in module
