+++ models_evaluation/evaluation_requirements.txt
@@ -1,0 +1,1 @@
+pycountry~=19.8.18
